Add a new scheme, and expose controls per scheme
