+++ src/lib/sidebar/PipelineSchemeForm.svelte
@@ -13,16 +13,15 @@
   import { gjSchemeCollection } from "stores";
   import type { SchemeData } from "types";
   import ATF4Type from "../forms/ATF4Type.svelte";
-  import { getArbitraryScheme } from "./scheme_data";
+
+  export let scheme_reference: string;
 
   let showModal = false;
   let scheme: SchemeData | null = null;
 
-  // This component is only created once, but gjScheme could be cleared
-  // completely multiple times. Set defaults anytime the modal is open.
+  // Lazily set defaults when the modal is opened the first time for a scheme
   $: if (showModal) {
-    // TODO Change the UI to plumb in the scheme_reference, when we manage multiple
-    scheme = getArbitraryScheme($gjSchemeCollection);
+    scheme = $gjSchemeCollection.schemes[scheme_reference];
     scheme.pipeline ||= {
       scheme_type: "",
       status: "",
